feat: Introduce AsyncFileOperations for enhanced asynchronous file handling and refactor related components
