+++ src/Handlers/File/FileHandler.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rcalicdan\FiberAsync\Handlers\AsyncOperations;
+namespace Rcalicdan\FiberAsync\Handlers\File;
 
 use Rcalicdan\FiberAsync\AsyncEventLoop;
 use Rcalicdan\FiberAsync\AsyncPromise;
