+++ src/Handlers/File/FileOperationHandler.php
@@ -1,7 +1,5 @@
 <?php
 
-// src/Handlers/File/FileOperationHandler.php
-
 namespace Rcalicdan\FiberAsync\Handlers\File;
 
 use Rcalicdan\FiberAsync\ValueObjects\FileOperation;
@@ -242,7 +240,7 @@ final readonly class FileOperationHandler
 
         $files = array_diff(scandir($dir), ['.', '..']);
         foreach ($files as $file) {
-            $path = $dir.DIRECTORY_SEPARATOR.$file;
+            $path = $dir . DIRECTORY_SEPARATOR . $file;
             if (is_dir($path)) {
                 $this->removeDirectoryRecursive($path);
             } else {
